requirements.txt
python-telegram-bot>=20.0
requests>=2.28.0
python-dotenv>=0.21.0
aiohttp>=3.8.3
newsapi-python>=0.2.6
pydantic>=1.10.0
loguru>=0.6.0 